fix(macos): address Copilot review comments
- Move tests to window_tests.rs following project convention
- Use AutoreleasePoolGuard for RAII pool management
- Use to_str() instead of to_string_lossy() for file paths
