+++ crates/warpui/src/platform/mac/window.rs
@@ -1691,60 +1691,5 @@ unsafe fn to_string(value: *mut Object) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::to_string;
-    use cocoa::base::{id, nil};
-    use cocoa::foundation::{NSAutoreleasePool, NSString};
-
-    unsafe fn nsstring(s: &str) -> id {
-        NSString::alloc(nil).init_str(s).autorelease()
-    }
-
-    #[test]
-    fn to_string_ascii() {
-        unsafe {
-            let pool = NSAutoreleasePool::new(nil);
-            assert_eq!(to_string(nsstring("hello world")), "hello world");
-            pool.drain();
-        }
-    }
-
-    #[test]
-    fn to_string_chinese() {
-        unsafe {
-            let pool = NSAutoreleasePool::new(nil);
-            assert_eq!(to_string(nsstring("中文")), "中文");
-            pool.drain();
-        }
-    }
-
-    #[test]
-    fn to_string_japanese() {
-        unsafe {
-            let pool = NSAutoreleasePool::new(nil);
-            assert_eq!(to_string(nsstring("日本語")), "日本語");
-            pool.drain();
-        }
-    }
-
-    #[test]
-    fn to_string_emoji() {
-        unsafe {
-            let pool = NSAutoreleasePool::new(nil);
-            assert_eq!(to_string(nsstring("🎉")), "🎉");
-            pool.drain();
-        }
-    }
-
-    #[test]
-    fn to_string_mixed_cjk() {
-        unsafe {
-            let pool = NSAutoreleasePool::new(nil);
-            assert_eq!(
-                to_string(nsstring("hello 中文 world 日本語 test")),
-                "hello 中文 world 日本語 test"
-            );
-            pool.drain();
-        }
-    }
-}
+#[path = "window_tests.rs"]
+mod tests;
